server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
pydantic==2.8.0
python-dotenv==1.0.1
